requirements.txt
asgiref==3.6.0
Django==4.1.5
django-ckeditor==6.5.1
django-js-asset==2.0.0
django-model-utils==4.3.1
fontawesomefree==5.15.4
Pillow==9.4.0
psycopg2==2.9.5
sqlparse==0.4.3
tzdata==2022.7
Unipath==1.1